backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0

# Environment variables
python-dotenv==1.0.0

# Multi-agent / LLM
crewai==0.63.2
langchain==0.2.0
langchain-groq==0.3.8
groq==0.3.0

# OCR / image handling
pytesseract==0.3.10
Pillow==10.1.0

# Database
sqlalchemy==2.0.23
databases==0.6.3  # optional async DB helper

# Reminders / scheduling / email
apscheduler==3.10.4
fastapi-mail==1.4.1

# File uploads
python-multipart==0.0.6
aiofiles==23.2.1

# Security
bcrypt==4.1.2

# HTTP requests
requests==2.31.0
